fix(sim): unusable UDS configs must not start, and every target must be reachable (#74 r2)
Five findings, all real, and two of them meant the feature was half-wired.

- **The Diagnostics panel could not reach any configured ECU.** It opened 0x7E0/0x7E8
  unconditionally, so every per-ECU target — including the ChassisECU I added to the demo in
  the same PR — was unreachable from the one UI that exists to reach it. The panel now lists
  the running targets and addresses the selected one.
- **A disabled ECU kept answering diagnostics.** The server loop followed only the channel's
  running flag, so unticking a node removed its frames and left its UDS server replying: a test
  simulating an offline ECU still found one, and the panel said it was off. The loop follows
  that node's enabled state now.
- **Invalid configurations were warned about and then spawned anyway** — the worst combination.
  Duplicate listeners both answered, and an entry with a missing address still made the list
  non-empty, which suppressed the working channel default in favour of a server that could
  never reply. `uds_nodes` now returns only usable servers; the rejected ones are still
  explained.
- **29-bit addressing was ignored.** Every channel opened standard, so a normal fixed address
  like 0x18DAF110 is masked by SocketCAN to 0x110 — a different and probably occupied address.
  Inferred from the ids, since nothing above 0x7FF can be 11-bit; a pair mixing the two formats
  is reported, because ISO-TP uses one format for both.
- **A DID above 0xFFFF was silently truncated**, so 0x1F190 became 0xF190 and overwrote or
  impersonated the VIN. `DidCfg.id` is u32 so validation can see the mistake, and the entry is
  dropped rather than narrowed.

Verified: v test modules/ 36/36, GUI compile-links, headless runner builds, diag_basic 6/6 and
diag_advanced 4/4 still pass.

+++ modules/project/project.v
@@ -109,7 +109,10 @@ pub mut:
 // bytes.
 pub struct DidCfg {
 pub mut:
-	id    u16
+	// u32, not u16: narrowing at parse time made 0x1F190 silently become 0xF190, which then
+	// masquerades as — or overwrites — a different configured DID. Kept wide so validation can
+	// see the mistake, and narrowed only once it is known to fit.
+	id    u32
 	text  string
 	bytes []u8
 }
@@ -525,7 +528,7 @@ fn parse_node(n yaml.Any) NodeCfg {
 		if ds := u.value_opt('dids') {
 			for d in ds.array() {
 				mut dc := DidCfg{
-					id:   u16(parse_id(d.value('id').str()))
+					id:   parse_id(d.value('id').str())
 					text: d.value('text').default_to('').string()
 				}
 				if bv := d.value_opt('bytes') {
